Fix type

+++ LOS_Driver/RC_Sender/Src/LOS_D_PPMChannelOut.cpp
@@ -68,7 +68,7 @@ uint32_t PPMChannelOut::getNextPPM()
         output_index_ = 0;
     }
 
-    uint8_t output = ppm_output_[output_index_++];
+    uint32_t output = ppm_output_[output_index_++];
     return output;
 }
